+++ src/constants/characterPersonalities.ts
