+++ src/constants/creatureAttributes.ts
@@ -1,4 +1,4 @@
-export const CREATURE_ATTRIBUTE: Record<string, string> = {
+export const CREATURE_ATTRIBUTES: Record<string, string> = {
   Insects: '곤충',
   Fish: '물고기',
   'Sea Creatures': '해양생물',
